* feature/SmartInfo:
  VISBO Zertifikat eingetragen
  Korrektur in awinImportProjectmitHrchy: es werden in Ressourcen und Termine die zwei Zellen in der ersten Spalte verbunden, damit das Einlesen dieser Daten korrekt abläuft. Meist sind diese Zellen verbunden, doch bei Änderungen im Steckbrief ein Fehlerquelle.
  Korrekturen am 2.1.17
  letzte Änderungen für Mahle Auslieferung
  finale Änderungen vor Auslieferung Mahle
  beil Laden von Projekten: es werden die bereits gezeigten Varianten im ShowPRojekte gelassen; alle anderen Varianten kommen in die Session in den NoShow Speicher Check Constellation Methode , wird aufgerufen wenn visboDebug = true Einstellungen auslieferfähig
  überall überprüft, dass currentConstellation aktualisiert wird
  nur noch speichern, was sich verändert hat smartInfo: UserLizenz geht bis 27.1.16 bei Varianten werden die CustomFields aus der Basis-Variate kopiert, wenn sie nicht vorhanden sind jede Phase bekommt jetzt eine leere Bewertung das Verschieben von Projekten auf der Projekt-Tafel geht jetzt auch nebeneinander frmProjPortfolioAdin beschleunigt: Varianten werden nicht mehr geckecht , erst bei Expand TrashCollections für Projekte und Constellations eingeführt
  Pfad für die .dll von MongoDBAccess, ProjectBoardBasic und ProjectBoardDefinitions
  Mahle Anpassungen Dez 2016
  ZwischenStand Auslieferung Mahle
  browserConstellations realisieren ...
  Methode Projekt-Varianten Namen aus der Datenbank lesen eingeführt, um das Lesen von DB wesentlich zu beschleunigen BuildTreeView und updateTreeView , um nach Filtern die TreeView entsprechend darzustellen
  Fehler behoben: beim Varianten anlegen wurden keine CustomFields übernommen ...
  Fehlerbehebung, die beim Testen für Auslieferung an Telair entdeckt wurden. Außerdem wird neuer ProjektSteckbrief benötigt und die ReportVorlagen für Telair mussten angepasst werden
  Verweise alle auf Microsoft Office Object Library 14 ,Microsoft Excel Object Library 14, Microsoft Powerpoint Object Library 14 gelenkt kein Microsoft.Interop Excel aus Visual Studio ...
  Aktueller Stand vor Änderung der Verweise (Excel Object Libarary ersetzt Interop.Excel)
  AutoReduce auch bei Summeneingabe eingeführt .. MS Project Add-In Änderung im Ribbon11: reportAuswahl.calledFrom = MS Project
  wichtige Änderung: in Ribbon1.vb von Visbo 1Click-PPT nämlich die Zuweisung, dass der Aufruf von frmReportProfil calledFrom = "MS Project" ist

+++ Projectboard/MongoDbAccess/Request.cs
@@ -36,7 +36,9 @@ namespace MongoDbAccess
         protected  IMongoDatabase Database;
         protected MongoServer Server;
         protected  IMongoCollection<clsProjektDB> CollectionProjects;
+        protected  IMongoCollection<clsProjektDB> CollectionTrashProjects;
         protected  IMongoCollection<clsConstellationDB> CollectionConstellations;
+        protected IMongoCollection<clsConstellationDB> CollectionTrashConstellations; 
         protected  IMongoCollection<clsDependenciesOfPDB> CollectionDependencies;
         protected  IMongoCollection<clsFilterDB> CollectionFilter;
         
@@ -74,7 +76,9 @@ namespace MongoDbAccess
             
                       
             CollectionProjects = Database.GetCollection<clsProjektDB>("projects");
+            CollectionTrashProjects = Database.GetCollection<clsProjektDB>("trashprojects");
             CollectionConstellations = Database.GetCollection<clsConstellationDB>("constellations");
+            CollectionTrashConstellations = Database.GetCollection<clsConstellationDB>("trashconstellations");
             CollectionDependencies = Database.GetCollection<clsDependenciesOfPDB>("dependencies");
             CollectionFilter = Database.GetCollection<clsFilterDB>("filters");
 
@@ -279,40 +283,56 @@ namespace MongoDbAccess
               
                                        
         }
-        //************************************/
-        public bool deleteProjectHistoryFromDB(string projectname, string variantName, DateTime storedEarliest, DateTime storedLatest)
-        {
 
+        public bool storeProjectToTrash(clsProjektDB projektDB)
+        {
             try
             {
-                storedLatest = storedLatest.ToUniversalTime();
-                storedEarliest = storedEarliest.ToUniversalTime();
-                string searchstr = Projekte.calcProjektKeyDB(projectname, variantName);
-                               
-
-                var dResult = CollectionProjects.DeleteMany<clsProjektDB>(p => (p.name == searchstr && p.timestamp >= storedEarliest && p.timestamp <= storedLatest));
-                if (dResult.DeletedCount > 0 )
-                    { return true; }
-                else
-                    { return false; }
                 
+                CollectionTrashProjects.InsertOne(projektDB);
+                  
+                return true;
             }
             catch (Exception)
             {
-                
                 return false;
             }
-            
-            // das folgende geht noch nicht , wer weiss warum ? 
-            ////CollectionProjects.DeleteMany<clsProjektDB>(query);
-            ////CollectionProjects.DeleteMany<clsProjektDB>(query);
-            
-            
-            // alt 2.x 
-            //var query = Query<clsProjektDB>
-            //         .Where(p => (p.name == searchstr && p.timestamp >= storedEarliest && p.timestamp <= storedLatest));
-            //return !CollectionProjects.Remove(query).HasLastErrorMessage;
         }
+        //************************************/
+        // darf nicht mehr verwendet werden, weil damit keine Speicherung der TimeStamps im Papierkorb verbunden ist ... 
+        ////public bool deleteProjectHistoryFromDB(string projectname, string variantName, DateTime storedEarliest, DateTime storedLatest)
+        ////{
+
+        ////    try
+        ////    {
+        ////        storedLatest = storedLatest.ToUniversalTime();
+        ////        storedEarliest = storedEarliest.ToUniversalTime();
+        ////        string searchstr = Projekte.calcProjektKeyDB(projectname, variantName);
+                               
+
+        ////        var dResult = CollectionProjects.DeleteMany<clsProjektDB>(p => (p.name == searchstr && p.timestamp >= storedEarliest && p.timestamp <= storedLatest));
+        ////        if (dResult.DeletedCount > 0 )
+        ////            { return true; }
+        ////        else
+        ////            { return false; }
+                
+        ////    }
+        ////    catch (Exception)
+        ////    {
+                
+        ////        return false;
+        ////    }
+            
+        ////    // das folgende geht noch nicht , wer weiss warum ? 
+        ////    ////CollectionProjects.DeleteMany<clsProjektDB>(query);
+        ////    ////CollectionProjects.DeleteMany<clsProjektDB>(query);
+            
+            
+        ////    // alt 2.x 
+        ////    //var query = Query<clsProjektDB>
+        ////    //         .Where(p => (p.name == searchstr && p.timestamp >= storedEarliest && p.timestamp <= storedLatest));
+        ////    //return !CollectionProjects.Remove(query).HasLastErrorMessage;
+        ////}
 
         //************************************/
         public bool deleteProjectTimestampFromDB(string projectname, string variantName, DateTime stored)
@@ -326,12 +346,42 @@ namespace MongoDbAccess
                 var query = Query<clsProjektDB>
                             .Where(p => (p.name == searchstr && p.timestamp == stored));
 
-                var dResult = CollectionProjects.DeleteOne<clsProjektDB>(p => (p.name == searchstr && p.timestamp == stored));
                 
-                if (dResult.DeletedCount > 0)
-                { return true; }
+                var sResult = CollectionProjects.Find<clsProjektDB>(p => (p.name == searchstr && p.timestamp == stored));
+                
+                if (sResult == null)
+                {
+                    return false;
+                }
                 else
-                { return false; }
+                {
+                    try
+                    {
+                        clsProjektDB projektDB = sResult.Single();
+                        if (storeProjectToTrash(projektDB))
+                        {
+                            // jetzt wird erst gelöscht 
+                            var dResult = CollectionProjects.DeleteOne<clsProjektDB>(p => (p.name == searchstr && p.timestamp == stored));
+
+                            if (dResult.DeletedCount > 0)
+                            { return true; }
+                            else
+                            { return false; }
+                        }
+                        else
+                        {
+                            return false;
+                        }
+
+                        
+                    }
+                    catch (Exception)
+                    {
+                        return false; 
+                    }
+                                      
+                }
+                
                
             }
             catch (Exception)
@@ -344,6 +394,62 @@ namespace MongoDbAccess
             // alt: 2.x 
             //return !CollectionProjects.Remove(query).HasLastErrorMessage;
         }
+        /// <summary>
+        /// liest alle vorkommenden Namen ProjektName#VariantenName aus der Datenbank , die zum Zeitpunkt storedLatest auch in der Datenbank existiert haben 
+        /// dabei wird ein übergebener Zeitraum berücksichtigt ... also nur Projekte, die auch im Zeitraum liegen ...
+        /// </summary>
+        /// <param name="zeitraumStart"></param>
+        /// <param name="zeitraumEnde"></param>
+        /// <param name="storedEarliest"></param>
+        /// <param name="storedatOrBefore"></param>
+        /// <returns></returns>
+        public SortedList<string, string> retrieveProjectVariantNamesFromDB(DateTime zeitraumStart, DateTime zeitraumEnde, DateTime storedatOrBefore)
+        {
+            var result = new SortedList<string, string>();
+
+            // in der Datenbank sind die Zeiten als Universal time gespeichert .. 
+            // deshalb muss hier umgerechnet werden 
+            storedatOrBefore = storedatOrBefore.ToUniversalTime();
+            
+            int startMonat = (int)DateAndTime.DateDiff(DateInterval.Month, Module1.StartofCalendar, zeitraumStart) + 1;
+            
+                
+            var prequery = CollectionProjects.AsQueryable<clsProjektDB>()
+                            .Where(c => c.startDate <= zeitraumEnde && c.endDate >= zeitraumStart && c.timestamp <= storedatOrBefore)
+                            .Select(c => c.name)
+                            .Distinct()
+                            .ToList();
+
+            foreach (string name in prequery)
+                {
+                                        
+                    try
+                    {
+
+                        if  (result.ContainsKey (name))  
+                        {
+                            // nichts tun 
+                        }
+                        else
+                        {
+                            result.Add(name, name);
+                        }
+                        
+
+                    }
+                    catch (Exception)
+                    {
+
+                        // nichts tun ...
+                    }
+
+
+                }
+          
+
+            return result;
+        }
+
 
         public SortedList<string, clsProjekt> retrieveProjectsFromDB(string projectname, string variantName, DateTime zeitraumStart, DateTime zeitraumEnde, DateTime storedEarliest, DateTime storedLatest, bool onlyLatest)
         {
@@ -455,7 +561,9 @@ namespace MongoDbAccess
 
             var prequery = CollectionProjects.AsQueryable<clsProjektDB>()
                             .Where(c => c.name.Contains(searchstr))
+                            .OrderBy(c => c.variantName)
                             .Select(c => c.variantName)
+                            .ToList()
                             .Distinct();
 
             foreach (string vName in prequery)
@@ -600,18 +708,61 @@ namespace MongoDbAccess
            
         }
 
+        public bool storeConstellationToTrash(clsConstellation c)
+        {
+            try
+            {
+                var cDB = new clsConstellationDB();
+                cDB.copyfrom(c);
+                cDB.Id = cDB.constellationName;
+
+                bool alreadyExisting = CollectionTrashConstellations.AsQueryable<clsConstellationDB>()
+                        .Any(p => p.constellationName == c.constellationName);
+
+                if (alreadyExisting)
+                {
+                    var filter = Builders<clsConstellationDB>.Filter.Eq("constellationName", c.constellationName);
+
+                    var rResult = CollectionTrashConstellations.ReplaceOne(filter, cDB);
+                    if (rResult.ModifiedCount > 0)
+                    { return true; }
+                    else
+                    { return false; }
+                }
+                else
+                {
+                    CollectionTrashConstellations.InsertOne(cDB);
+                    return true;
+                }
+
+            }
+            catch (Exception)
+            {
+
+                return false;
+            }
+        }
         public bool removeConstellationFromDB(clsConstellation c)
         {
             
           try 
 	        {	        
 		    // neu 3.0 
-            var dResult = CollectionConstellations.DeleteOne<clsConstellationDB>(p => (p.Id == c.constellationName));
-              if (dResult.DeletedCount >0 )
-              { return true; }
+
+              if (storeConstellationToTrash (c))
+              {
+                  var dResult = CollectionConstellations.DeleteOne<clsConstellationDB>(p => (p.Id == c.constellationName));
+                  if (dResult.DeletedCount > 0)
+                  { return true; }
+                  else
+                  { return false; }
+              }
               else
-              { return false; }
+              {
+                  return false;
+              }
             
+           
             
             // alt 2.x
             //var query = Query<clsConstellationDB>.EQ(e => e.Id, c.constellationName);
@@ -693,7 +844,7 @@ namespace MongoDbAccess
                         {
                             var c = new clsConstellation();
                             cDB.copyto(ref c);
-                            int a = c.rename(oldName, newName);
+                            int a = c.renameProject(oldName, newName);
 
                            if (a>0)
                            {
